Update to 0.1.1
Updated to use Khan's algorithm instead of recursion since testing at work with a lot of scripts used 20GB of memory then crashed, lol.

Also updated the regex and remove script types since it's not really needed - might add this back at a later date if I find a use for it.